# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Constituent Touchpoint Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaign-types:
    get:
      operationId: listCampaignTypes
      tags: [campaignTypes]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CampaignType" }
    post:
      operationId: createCampaignType
      tags: [campaignTypes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CampaignTypeInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignType" }
  /campaign-types/{id}:
    patch:
      operationId: updateCampaignType
      tags: [campaignTypes]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CampaignTypeUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignType" }

  # ───────── Channels
  /channels:
    get:
      operationId: listChannels
      tags: [channels]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Channel" }
    post:
      operationId: createChannel
      tags: [channels]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChannelInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Channel" }
  /channels/{id}:
    patch:
      operationId: updateChannel
      tags: [channels]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChannelUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Channel" }

  # ───────── Campaigns
...
  /campaigns/{id}/seeds:
    get:
      operationId: listSeeds
      tags: [seeds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SeedGroup" }
    post:
      operationId: createSeedGroup
      tags: [seeds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SeedGroupInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SeedGroup" }
  /campaigns/{id}/seeds/{seedId}:
    delete:
      operationId: deleteSeedGroup
      tags: [seeds]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: seedId, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }

  # ───────── Preview / Export
  /campaigns/{id}/preview:
    get:
      operationId: getCampaignPreview
      tags: [exports]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignPreview" }
  /campaigns/{id}/finalize:
    post:
      operationId: finalizeCampaign
      tags: [exports]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
  /campaigns/{id}/export:
    post:
      operationId: exportCampaign
      tags: [exports]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExportResult" }

  # ───────── Donor lookup
...
components:
  schemas:
...
    CampaignType:
      type: object
      required: [id, name, active, systemDefault]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        active: { type: boolean }
        systemDefault: { type: boolean }
    CampaignTypeInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
    CampaignTypeUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        active: { type: boolean }
    Channel:
      type: object
      required: [id, name, active, systemDefault]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        active: { type: boolean }
        systemDefault: { type: boolean }
    ChannelInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
    ChannelUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        active: { type: boolean }
...
    Campaign:
      type: object
      required: [id, name, status, submittedByUserId, submittedByName, createdAt, campaignTypes]
      properties:
        id: { type: integer }
        name: { type: string }
        status: { type: string }
        owningUnit: { type: ["string", "null"] }
        submittedByUserId: { type: integer }
        submittedByName: { type: string }
        intendedSendStartDate: { type: ["string", "null"], format: date }
        audienceDescription: { type: ["string", "null"] }
        salesforceCampaignId: { type: ["string", "null"] }
        internalNotes: { type: ["string", "null"] }
        originalRowCount: { type: integer }
        blankRowCount: { type: integer }
        validIdCount: { type: integer }
        uniqueIdCount: { type: integer }
        duplicateIdCount: { type: integer }
        rejectedIdCount: { type: integer }
        extraColumnsIgnored: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        exportedAt: { type: ["string", "null"], format: date-time }
        archivedAt: { type: ["string", "null"], format: date-time }
        voidedAt: { type: ["string", "null"], format: date-time }
        campaignTypes:
          type: array
          items: { $ref: "#/components/schemas/CampaignType" }
...
    SeedGroup:
      type: object
      required: [id, campaignId, scope, seedCount, createdAt]
      properties:
        id: { type: integer }
        campaignId: { type: integer }
        scope: { type: string, enum: [all, channel, touch] }
        channelId: { type: ["integer", "null"] }
        touchId: { type: ["integer", "null"] }
        seedCount: { type: integer }
        createdAt: { type: string, format: date-time }
    SeedGroupInput:
      type: object
      required: [scope]
      properties:
        scope: { type: string, enum: [all, channel, touch] }
        channelId: { type: integer }
        touchId: { type: integer }
        rawText: { type: string }
        googleSheetUrl: { type: string }
    CampaignPreview:
      type: object
      required: [campaignId, audienceUnique, totalPlannedTouchpointsBefore, totalPlannedTouchpointsAfter, perTouch]
      properties:
        campaignId: { type: integer }
        audienceUnique: { type: integer }
        thresholdFlaggedDonors: { type: integer }
        manuallySuppressedDonors: { type: integer }
        totalSeedIds: { type: integer }
        totalPlannedTouchpointsBefore: { type: integer }
        totalPlannedTouchpointsAfter: { type: integer }
        perTouch:
          type: array
          items: { $ref: "#/components/schemas/PerTouchPreview" }
    PerTouchPreview:
      type: object
      required: [touchId, touchName, channelLabel, campaignTypeLabel, sendDate, eligibleCount, suppressedCount, seedCount, totalRowsInExport, fileName]
      properties:
        touchId: { type: integer }
        touchName: { type: string }
        channelLabel: { type: string }
        campaignTypeLabel: { type: string }
        sendDate: { type: string, format: date }
        eligibleCount: { type: integer }
        suppressedCount: { type: integer }
        seedCount: { type: integer }
        totalRowsInExport: { type: integer }
        fileName: { type: string }
    ExportResult:
      type: object
      required: [campaignId, status, exportedAt, files]
      properties:
        campaignId: { type: integer }
        status: { type: string }
        exportedAt: { type: string, format: date-time }
        files:
          type: array
          items:
            type: object
            required: [touchId, fileName, rowCount, downloadUrl]
            properties:
              touchId: { type: integer }
              fileName: { type: string }
              rowCount: { type: integer }
              seedCount: { type: integer }
              suppressedCount: { type: integer }
              downloadUrl: { type: string }
...